Switch from maven-assembly to shading plugin to fix issues with log4j
log4j generates a cache on compilation that contains all plugins, if multiple
plugins are specified, shading will go wrong and override all but one cache
to prevent that, we switch to shading plugin and exclude the cache file from
the shaded jar. this wil no really resolve the problem but make log4j build
the cache on runtime. At the cost some performance that is...

Closes #8